Many changes - Work done more ingrained in the game, cards can be dragged
WorkDone has been a metric that did not do anything. Now we are integrating it into a second win condition, as well as an overarching goal that has to be reached at the end of the week. Items can combo into generating more work done. Will add encounters that consist of meeting a certain threshold. Also added the Clock system back, currently stuck on 8 turns but will depend on the length of the meeting. Each turn is 1 hour so if the meeting is 3 hours we will only have 3 turns to complete it.

Cards can now be dragged and will follow the mouse instead of being static in hand.